Sitzverteilung simplification + now takes wahlJahr prop (#28)
